backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.37
alembic==1.14.0
httpx==0.27.2
feedparser==6.0.11
python-dotenv==1.0.1
pydantic-settings==2.6.1
